[ty] Move some `pub` functions in `ty_python_core` to be `pub(crate)` functions in `ty_python_semantic`
