+++ crates/ty_python_semantic/src/types/infer/builder/post_inference/static_class.rs
@@ -9,6 +9,7 @@ use ruff_python_ast as ast;
 use ruff_text_size::{Ranged, TextRange, TextSize};
 use rustc_hash::FxHashMap;
 
+use crate::attribute_assignments;
 use crate::{
     TypeQualifiers,
     diagnostic::format_enumeration,
@@ -49,9 +50,7 @@ use crate::{
         visitor::find_over_type,
     },
 };
-use ty_python_core::{
-    SemanticIndex, attribute_assignments, definition::DefinitionKind, scope::ScopeId,
-};
+use ty_python_core::{SemanticIndex, definition::DefinitionKind, scope::ScopeId};
 
 /// Iterate over all static class definitions (created using `class` statements) to check that
 /// the definition will not cause an exception to be raised at runtime. This needs to be done
